% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposit_funds.R
\name{deposit_funds}
\alias{deposit_funds}
\title{Function to Deposit Funds via Sign In with Coinbase API}
\usage{
deposit_funds(
  amount = "1.00",
  currency = "USD",
  payment_method = "",
  commit = "true",
  deposit_account_id = ""
)
}
\arguments{
\item{amount}{Deposit amount}

\item{currency}{Currency in which \code{amount} is denominated.}

\item{payment_method}{ID of payment method to be used for the deposit. List
Payment Methods: GET /payment-methods}

\item{commit}{A boolean, if \code{true} the deposit is to be immediately
executed. If \code{false}, the the deposit will not be
immediately executed. In this case, use \code{deposit_commit}
to execute.
\code{base_size} can be expressed}
}
\value{
A string to pass as payload/body for the Market Immediate-or_Cancel
order
}
\description{
Remember,
Base currency is the first listed (e.g., \code{BTC} in \code{BTC-USD})
Quote currency is the last listed (e.g., \code{USD} in \code{BTC-USD})
}
\details{
The documentation appears to be wrong, market orders must be specified in
quote currencies
}
