% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_SWCB_API_keys.R
\name{interact_SWCB_API_keys}
\alias{interact_SWCB_API_keys}
\title{Interact with the Coinbase Advanced Trading API}
\usage{
interact_SWCB_API_keys(
  api_key = "",
  secret_key = "",
  method = c("GET", "POST"),
  reqPath = "",
  query = "",
  body = ""
)
}
\arguments{
\item{api_key}{The api key assigned to the Advanced Trading Wallet}

\item{secret_key}{The secret key assigned to the Advanced Trading Wallet}

\item{method}{Select whether to send a \code{GET} or \code{POST} request}

\item{reqPath}{The request path, can omit \code{"/v3/api/brokerage"}}

\item{query}{The query items to add to the end of the url}

\item{body}{The body/payload to pass to the API, usually used with method \code{"POST"}}
}
\value{
Returns the content from the API request
}
\description{
This a function that allows to use "GET" or "POST" methods to interact with
Coinbase API
}
\details{
Added the "query" after it was loaded
}
